Factor input handling into it's own module
